# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Student Management API",
    "description": "API для управління навчальним процесом"
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Введіть токен у форматі: Bearer <Ваш_токен>"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/students/register": {
      "post": {
        "tags": ["Students"],
        "summary": "Реєстрація нового студента",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/StudentRegister" }
          }
        ],
        "responses": {
          "201": { "description": "Студент успішно зареєстрований" },
          "400": { "description": "Помилка реєстрації" }
        }
      }
    },
    "/students/login": {
      "post": {
        "tags": ["Students"],
        "summary": "Вхід для студента",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Login" }
          }
        ],
        "responses": {
          "200": { "description": "Успішний вхід" },
          "404": { "description": "Студента не знайдено" }
        }
      }
    },
    "/students": {
      "get": {
        "tags": ["Students"],
        "summary": "Отримати список всіх студентів",
        "description": "Повертає список усіх зареєстрованих студентів.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Успішний запит. Повертається список студентів.",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Student" }
            }
          },
          "500": {
            "description": "Помилка сервера."
          }
        }
      }
    },
    "/students/{id}": {
      "put": {
        "tags": ["Students"],
        "summary": "Оновлення даних студента",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID студента для оновлення"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/StudentUpdate" }
          }
        ],
        "responses": {
          "200": { "description": "Дані успішно оновлено" },
          "400": { "description": "Некоректні дані" },
          "404": { "description": "Студента не знайдено" }
        }
      },
      "delete": {
        "tags": ["Students"],
        "summary": "Видалити студента",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID студента, якого потрібно видалити"
          }
        ],
        "responses": {
          "200": { "description": "Студента  успішно видалено" },
          "404": { "description": "Студента не знайдено" },
          "500": { "description": "Помилка сервера" }
        }
      }
    },
    "/students/class/{class}": {
      "get": {
        "tags": ["Students"],
        "summary": "Пошук студентів за класом",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "class",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Список студентів" }
        }
      }
    },
    "/students/grades/export": {
      "get": {
        "tags": ["Students"],
        "security": [{ "BearerAuth": [] }],
        "summary": "Генерація звіту про оцінки студента у форматі Word",
        "produces": ["application/vnd.openxmlformats-officedocument.wordprocessingml.document"],
        "responses": {
          "200": {
            "description": "Файл звіту у форматі Word",
            "content": {
              "application/vnd.openxmlformats-officedocument.wordprocessingml.document": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": { "description": "Студента не знайдено" },
          "500": { "description": "Помилка генерації файлу" }
        }
      }
    },
...
    "/teachers/student": {
      "get": {
        "tags": ["Teachers"],
        "summary": "Отримати список вчителів авторизованого учня",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Список вчителів",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Teacher" }
            }
          },
          "404": { "description": "Студента не знайдено" },
          "500": { "description": "Помилка сервера" }
        }
      }
    },
...
    "/subjects/student/{studentId}": {
      "get": {
        "tags": ["Subjects"],
        "summary": "Отримати список предметів конкретного студента",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "studentId", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Список предметів" },
          "404": { "description": "Студента не знайдено" }
        }
      }
    },
...
    "/grades/{id}": {
      "delete": {
        "tags": ["Grades"],
        "summary": "Видалити оцінку у студента",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id", "in": "path", "required": true, "type": "string"
          },
          { "name": "studentId", "in": "query", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Оцінку видалено" },
          "403": { "description": "Немає доступу до цієї оцінки" },
          "404": { "description": "Оцінку не знайдено" }
        }
      },
      "get": {
        "tags": ["Grades"],
        "summary": "Отримати оцінку за ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Дані оцінки" },
          "404": { "description": "Оцінку не знайдено" }
        }
      },
      "put": {
        "tags": ["Grades"],
        "summary": "Оновити оцінку",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/GradeUpdate" }
          }
        ],
        "responses": {
          "200": { "description": "Оцінку оновлено" },
          "403": { "description": "Немає доступу до цієї оцінки" },
          "404": { "description": "Оцінку не знайдено" }
        }
      }
    },
    "/grades": {
      "get": {
        "tags": ["Grades"],
        "summary": "Отримати список оцінок авторизованого студента",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Список оцінок" },
          "404": { "description": "Учня не знайдено" }
        }
      }
    },
    "/grades/student/{studentId}": {
      "get": {
        "tags": ["Grades"],
        "summary": "Отримати список оцінок конкретного студента",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "studentId", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Список оцінок" },
          "404": { "description": "Учня не знайдено" }
        }
      }
    },
    "/schedule": {
      "get": {
        "tags": ["Schedule"],
        "summary": "Отримати розклад авторизованого студента",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Розклад студента" },
          "404": { "description": "Студента не знайдено" }
        }
      }
    },
    "/schedule/student/{studentId}": {
      "get": {
        "tags": ["Schedule"],
        "summary": "Отримати розклад конкретного студента",
        "parameters": [
          { "name": "studentId", "in": "path", "required": true, "type": "string" }
        ],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Розклад студента" },
          "404": { "description": "Студента не знайдено" }
        }
      }
    },
    "/schedule/{id}": {
      "get": {
        "tags": ["Schedule"],
        "summary": "Отримати заняття з розкладу за ID",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Дані заняття" },
          "403": { "description": "Немає доступу до цього заняття" },
          "404": { "description": "Заняття не знайдено" }
        }
      },
      "put": {
        "tags": ["Schedule"],
        "summary": "Оновити заняття",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/ScheduleUpdate" }
          }
        ],
        "responses": {
          "200": { "description": "Заняття оновлено" },
          "403": { "description": "Немає доступу до цього заняття" },
          "404": { "description": "Заняття не знайдено" }
        }
      },
      "delete": {
        "tags": ["Schedule"],
        "summary": "Видалити заняття з розкладу",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          { "name": "studentId", "in": "query", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "Заняття видалено" },
          "403": { "description": "Немає доступу до цього заняття" },
          "404": { "description": "Заняття не знайдено" }
        }
      }
    },
    "/schedule/add": {
      "post": {
        "tags": ["Schedule"],
        "summary": "Додати заняття до розкладу студента",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/Schedule" }
          }
        ],
        "responses": {
          "201": { "description": "Заняття додано" },
          "400": { "description": "Помилка додавання заняття" }
        }
      }
    },
...
  },
  "definitions": {
    "StudentRegister": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "password": { "type": "string" },
        "surname": { "type": "string" },
        "name": { "type": "string" },
        "dateOfBirth": { "type": "string", "format": "date" },
        "studentClass": { "type": "string" }
      },
      "required": ["email", "password", "surname", "name", "dateOfBirth", "studentClass"]
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": { "type": "string" },
        "password": { "type": "string" }
      }
    },
    "Student": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "email": { "type": "string" },
        "surname": { "type": "string" },
        "name": { "type": "string" },
        "dateOfBirth": { "type": "string", "format": "date" },
        "class": { "type": "string" },
        "subjects": {
          "type": "array",
          "items": { "type": "string" }
        },
        "teachers": {
          "type": "array",
          "items": { "type": "string" }
        },
        "grades": {
          "type": "array",
          "items": { "type": "string" }
        },
        "schedule": {
          "type": "array",
          "items": { "type": "string" }
        }
      }
    },
    "StudentUpdate": {
      "type": "object",
      "properties": {
        "surname": { "type": "string" },
        "name": { "type": "string" },
        "dateOfBirth": { "type": "string", "format": "date" },
        "class": { "type": "string" }
      }
    },
    "Teacher": {
      "type": "object",
      "properties": {
        "surname": { "type": "string" },
        "name": { "type": "string" },
        "fatherName": {"type": "string"}
      }
    },
...
    "GradeUpdate": {
      "type": "object",
      "properties": {
        "subjectId": { "type": "string" },
        "grade": { "type": "number" },
        "date": { "type": "string", "format": "date" }
      }
    },
    "Schedule": {
      "type": "object",
      "properties": {
        "studentId": { "type": "string" },
        "subjectId": { "type": "string" },
        "dayOfWeek": { "type": "number" },
        "lessonNumber": { "type": "number" }
      }
    },
    "ScheduleUpdate": {
      "type": "object",
      "properties": {
        "dayOfWeek": { "type": "number" },
        "lessonNumber": { "type": "number" },
        "subjectId": { "type": "string" }
      }
    },
...
  }
}